Fixing test: we need new approach to testing
With Notify and Processing async servers, we need to know how to run the
tests. Some tests do not want or expect the notify to kick in after some
action, to not run all the side effects. The Notify and Processing
servers are not started for test by default and must be started with
`@moduletag start: [:notify, :processing]` or `@tag start: [:notify, :processing]`