Solve upsertTargets transaction timeout
This transaction is complex, does upsert of each given target and then
can remove all others - but delets them 1 by 1 because a target might
block due to it having messages. Then we return an exact id of offending
target. This would not work when batch deleting. The transaction was
optimized to do less reads and less deletes (rely on cascade from
targets to target_emails) and has a custom timeout of 2 minutes now to
accomodate for this.